Add Tree Traversal in Emojicode
This is a combination of 5 commits.
This is the 1st commit message:

Add Tree Traversal in Emojicode

This is the 2nd commit message:

Remove unnecessary 🤜🤛.

This was suggested by thbwd:
"No need to wrap 🤜children_count🤛 into 🤜🤛actually."
(https://github.com/algorithm-archivists/algorithm-archive/pull/
471#discussion_r223291032)

This is the 3rd commit message:

Extract shared logic into a method.

This was suggested by thbwd:
"This is shared logic. I’d recommend you extract it into a method."
(https://github.com/algorithm-archivists/algorithm-archive/pull/
471#discussion_r223291140)

This is the 4th commit message:

Use documentation comments.

This was suggested by thbwd:
"Shouldn’t these be documentation comments? https://www.emojicode.org/
docs/reference/documentation.html"
(https://github.com/algorithm-archivists/algorithm-archive/pull/
471#discussion_r223291281)

This is the 5th commit message:

Add error as a potential return value for DFS RI Binary

To be able to return an error and to apply suggestion by thbwd
(https://github.com/algorithm-archivists/algorithm-archive/pull/
471#discussion_r223597592) the following changes were made:
- Declare an enum to hold the error value.
- Rewrite the Depth-First Search Recursive Inorder Binary method to
return an optional holding the enum and adjust the returns according
to that.
- Call the rewritten method and check, if the returned optional holds
error. If it does: Print out the error message defined in the enum.

Signed-off-by: Julian Schacher <jspp@posteo.net>