Fix amibguous sentence in Scope-section.
The first sentence of the Scope-section didn't clearly indicate that the
"when an individual is representing the project or its community"-part
only applies to the second half of the sentence ("in public spaces") and
not the whole sentence.
Therefore change the layout of the sentence to remove all ambiguity.

For more info, see:
https://github.com/ContributorCovenant/contributor_covenant/issues/659
https://github.com/ContributorCovenant/contributor_covenant/pull/660